Split controller constructor off from reconciler logic.
