+++ pkg/reconciler/buildtemplate/buildtemplate.go
@@ -30,21 +30,16 @@ import (
 	"github.com/knative/pkg/controller"
 	"github.com/knative/pkg/kmeta"
 	"github.com/knative/pkg/logging"
-	"github.com/knative/pkg/logging/logkey"
 
 	"github.com/knative/build/pkg/apis/build/v1alpha1"
 	clientset "github.com/knative/build/pkg/client/clientset/versioned"
 	buildscheme "github.com/knative/build/pkg/client/clientset/versioned/scheme"
-	informers "github.com/knative/build/pkg/client/informers/externalversions/build/v1alpha1"
 	listers "github.com/knative/build/pkg/client/listers/build/v1alpha1"
 	"github.com/knative/build/pkg/reconciler/buildtemplate/resources"
 	caching "github.com/knative/caching/pkg/apis/caching/v1alpha1"
 	cachingclientset "github.com/knative/caching/pkg/client/clientset/versioned"
-	cachinginformers "github.com/knative/caching/pkg/client/informers/externalversions/caching/v1alpha1"
 	cachinglisters "github.com/knative/caching/pkg/client/listers/caching/v1alpha1"
 )
-
-const controllerAgentName = "buildtemplate-controller"
 
 // Reconciler is the controller.Reconciler implementation for BuildTemplates resources
 type Reconciler struct {
@@ -73,41 +68,6 @@ func init() {
 	// Add buildtemplate-controller types to the default Kubernetes Scheme so Events can be
 	// logged for buildtemplate-controller types.
 	buildscheme.AddToScheme(scheme.Scheme)
-}
-
-// NewController returns a new build template controller
-func NewController(
-	logger *zap.SugaredLogger,
-	kubeclientset kubernetes.Interface,
-	buildclientset clientset.Interface,
-	cachingclientset cachingclientset.Interface,
-	buildTemplateInformer informers.BuildTemplateInformer,
-	imageInformer cachinginformers.ImageInformer,
-) *controller.Impl {
-
-	// Enrich the logs with controller name
-	logger = logger.Named(controllerAgentName).With(zap.String(logkey.ControllerType, controllerAgentName))
-
-	r := &Reconciler{
-		kubeclientset:        kubeclientset,
-		buildclientset:       buildclientset,
-		cachingclientset:     cachingclientset,
-		buildTemplatesLister: buildTemplateInformer.Lister(),
-		imagesLister:         imageInformer.Lister(),
-		Logger:               logger,
-	}
-	impl := controller.NewImpl(r, logger, "BuildTemplates")
-
-	logger.Info("Setting up event handlers")
-	// Set up an event handler for when BuildTemplate resources change
-	buildTemplateInformer.Informer().AddEventHandler(controller.HandleAll(impl.Enqueue))
-
-	imageInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("BuildTemplate")),
-		Handler:    controller.HandleAll(impl.EnqueueControllerOf),
-	})
-
-	return impl
 }
 
 // Reconcile implements controller.Reconciler
